src/ask_finance/requirements.txt
# Core document processing
docling>=1.0.0
pandas>=1.5.0
numpy>=1.24.0
Pillow>=9.0.0

# Vector embeddings and search
sentence-transformers>=2.2.0
chromadb>=0.4.0
faiss-cpu>=1.7.0

# Language models and chains
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-core>=0.1.0
langgraph>=0.0.20
openai>=1.0.0

# Structured output parsing
pydantic>=2.0.0

# SQL and database
sqlite3  # Built-in with Python
sqlparse>=0.4.0

# Async support
asyncio  # Built-in with Python
aiofiles>=0.8.0

# Utility libraries
python-dotenv>=0.19.0
tqdm>=4.64.0
requests>=2.28.0

# Optional: For advanced image processing
transformers>=4.20.0
torch>=1.12.0  # For image models

# Development and testing
pytest>=7.0.0
pytest-asyncio>=0.20.0
black>=22.0.0
flake8>=5.0.0
mypy>=1.0.0

# Jupyter notebook support (optional)
jupyter>=1.0.0
ipykernel>=6.0.0